Splits out email functionality to separate file for cleaner view
